backend/requirements.txt
Flask==2.1.2
psycopg2-binary==2.9.1
requests==2.26.0
python-dotenv==0.19.1
Flask-Cors==4.0.0

